Rename hooks folder
