+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import CommentProvider from './hooks/CommentProvider';
+import CommentProvider from './providers/CommentProvider';
 
 ReactDOM.render(
   <React.StrictMode>
